HP가 왜 0일가요

+++ zelda_triforce/zelda_triforce/Player.h
@@ -12,7 +12,7 @@ private:
 
 	//╫╨ех
 	float _hp;
-	float _mp;
+	int _mp;
 	float _speed;
 
 	int _moneyCount;
